+++ mar11/forktest.c
@@ -1,5 +1,6 @@
 #include <errno.h>
 #include <fcntl.h>
+#include <inttypes.h>
 #include <pthread.h>
 #include <stdio.h>
 #include <stdlib.h>
@@ -50,7 +51,7 @@ main(int argc, char *argv[]) {
 			fprintf(stderr, "%s: Short read expected %d got %zd\n", me, BUFSIZE, nbytes);
 		}
 		/* Call lseek to get current offset. */
-		printf("%s: Offset is %ld\n", me, lseek(fd, 0, SEEK_CUR));
+		printf("%s: Offset is %" PRIuMAX "\n", me, (uintmax_t)lseek(fd, 0, SEEK_CUR));
 		sleep(1);
 	}
 	if (close(fd) != 0) {
